Fix some print format warnings
https://bugzilla.gnome.org/show_bug.cgi?id=734736

+++ gtk/gtkmountoperation-stub.c
@@ -60,7 +60,7 @@ _gtk_mount_operation_kill_process (GPid      pid,
                G_IO_ERROR,
                G_IO_ERROR_NOT_SUPPORTED,
                _("Cannot kill process with PID %d. Operation is not implemented."),
-               pid);
+               (int) pid);
   return FALSE;
 }
